requirements.txt
# Wildlife Risk Assessment System - Requirements
# Compatible with Python 3.13

# Core Backend Dependencies
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# Computer Vision and AI
opencv-python>=4.8.0
ultralytics>=8.3.0
torch>=2.6.0
torchvision>=0.21.0
numpy>=1.24.0
pillow>=10.0.0

# Alert Services
twilio>=8.10.0
python-dotenv>=1.0.0

# Additional utilities
requests>=2.31.0
python-dateutil>=2.8.0
pygame>=2.5.0

# Note: mediapipe and cvzone not yet compatible with Python 3.13
# If you need facial recognition, use Python 3.11 or earlier

# Optional: For GPU acceleration (if available)
# tensorflow>=2.15.0
# keras>=2.15.0

# Optional: For advanced AI models
# transformers>=4.35.0
# huggingface-hub>=0.19.4

# Testing (optional)
# pytest>=7.4.3
# pytest-flask>=1.3.0
